+++ src/main/java/team/jit/technicalinterviewdemo/book/BookRepository.java
@@ -1,9 +1,23 @@
 package team.jit.technicalinterviewdemo.book;
 
+import java.util.Optional;
+
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
+import org.springframework.data.jpa.repository.EntityGraph;
+import org.springframework.data.jpa.domain.Specification;
 
 public interface BookRepository extends JpaRepository<Book, Long>, JpaSpecificationExecutor<Book> {
+
+    @Override
+    @EntityGraph(attributePaths = "categories")
+    Optional<Book> findById(Long id);
+
+    @Override
+    @EntityGraph(attributePaths = "categories")
+    Page<Book> findAll(Specification<Book> spec, Pageable pageable);
 
     boolean existsByIsbn(String isbn);
 }